requirements.txt
alabaster==0.7.6
amqp==1.4.6
anyjson==0.3.3
Babel==2.0
billiard==3.3.0.20
characteristic==14.3.0
coverage==3.7.1
cryptography==1.0
cssselect==0.9.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-autoslug==1.8.0
django-jet==0.0.9
django-rest-swagger==0.3.4
django-toolbelt==0.0.1
djangorestframework==3.1.3
docutils==0.12
factory-boy==2.5.2
flake8==2.4.1
gunicorn==19.3.0
idna==2.0
Jinja2==2.8
kombu==3.0.26
MarkupSafe==0.23
mccabe==0.3.1
mock==1.3.0
msgpack-python==0.4.6
parsel==1.0.1
pbr==1.6.0
pep8==1.5.7
pluggy==0.3.0
psycopg2==2.6.1
py==1.4.30
pyasn1==0.1.8
pyasn1-modules==0.0.7
pycparser==2.14
PyDispatcher==2.0.5
pyflakes==0.8.1
Pygments==2.0.2
PyLinkedinAPI==0.1.2
pyOpenSSL==0.15.1
pytz==2015.4
requests==2.7.0
Scrapy==1.1.0.dev1
service-identity==14.0.0
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
PyYAML==3.11
static3==0.6.1
tox==2.1.1
virtualenv==13.1.2
wheel==0.24.0
zope.interface==4.1.2
